modules/infrastructure/persistence/mysql/modules_mysql_repository.go
/*
 * File: modules_mysql_repository.go
 * Author: bengie
 * Copyright: 2023, Smart Cities Peru.
 * License: MIT
 *
 * Purpose:
 * Repository for modules.
 *
 * Last Modified: 2023-11-10
 */

package mysql

import (
	"time"

	smartClock "gitlab.smartcitiesperu.com/smartone/api-shared/clock"

	moduleDomain "gitlab.smartcitiesperu.com/smartone/api-core/modules/domain"
	errDomain "gitlab.smartcitiesperu.com/smartone/api-shared/error-core/domain"
)

type modulesMySQLRepo struct {
	clock   smartClock.Clock
	timeout time.Duration
	err     *errDomain.SmartError
}

func NewModulesRepository(
	clock smartClock.Clock,
	mongoTimeout int,
) moduleDomain.ModuleRepository {
	rep := &modulesMySQLRepo{
		clock:   clock,
		timeout: time.Duration(mongoTimeout) * time.Second,
		err:     errDomain.NewErr().SetLayer(errDomain.Infra),
	}
	return rep
}
